oe-init-build-env now honors $1 if supplied, else BUILDENV.
Previously, both $1 and BUILDENV were ignored.

There is a bug in the evb-ast2500 build, the recommended workaround
for which is to run:

    TEMPLATECONF=meta-evb/meta-evb-aspeed/meta-evb-ast2500/conf source openbmc-env

instead of simply running:

    source setup ${MACHINE} ${BUILDDIR}

Recently, a change was merged that requires the build dir not to be a
subdir of /usr.  Glinux machines currently place home directories
under /usr, so BUILDDIR has to function correctly.  This change has
the side effect of making openbmc-env honor the BUILDDIR environment
variable.

Signed-off-by: Dan Nussbaum <dansn@google.com>
Change-Id: I7be405f5876f4fa1ba7cab4d464201db6a9b092e