meta-rust demo works for rust-hello, spitransport-tool and otpilot-tool.
All three are in a (new) apps directory, which is a subdir of

    meta-evb/meta-evb-aspeed/meta-evb-ast2500/recipes-phosphor

(also new).

In order to make this work, I had to add to BBFILES entries in

    meta-evb/meta-evb-aspeed/meta-evb-ast2500/conf/layer.conf

TODO: Currently, spitransport-tool.bb and otpilot-tool.bb have to list
their own rust package dependencies.  This should not be necessary --
there ought to be a way to make use of dependencies listed in
tock-on-titan.git:/cargo-registry.  `cargo bitbake` generates some of
the needed dependencies, but not all of them, and as of 2021-05-09, I
have not been able to dig up a way to make it work completely, so the
process of filling them out is iterative and somewhat laborious.

Signed-off-by: Dan Nussbaum <dansn@google.com>
Change-Id: I73cbdb415668433b790f2b04ee95ad22d6d6a44d